Revert "Renamed fisher_vector.py"
This reverts commit 7bfbe3d008c3e3863708d38107c2e2488dedd5e3.
