+++ skimage/feature/fisher_vector.py
